Resolve corrupt ouput for larg data
ncompress -d does not throw corrupt input for larger files compressed
with this program. I figured the problem would be related to the 9 bit
limit we had in the program which was indeed the source of trouble. But
now i don't have any idea why it DOES work. It really shouldn't! The '+
2' in the main.c:100 is resolving the problem. If you make it '+ 1' you
will get wrong decompression at some files at the last characters (but
no error) and if you make it '+ 0' you will be faced with corrupt input
error from ncompress -d. I just gotta dig a little deeper to understand
why